feat: summary table (#95)
* feat: summary table

* feat: concurrent update summary

* fix: ci lint

* fix: create index

* fix: update group summary

* fix: group summary index name

* fix: batch file

* feat: dashboard from summary

* fix: group summarys index

* fix: sum request count

* feat: char data default from summary

* fix: on conflict requires inference specification or constraint name

* fix: column reference is ambiguous

+++ relay/adaptor/ali/rerank.go
@@ -23,7 +23,7 @@ type RerankOutput struct {
 	Results []*model.RerankResult `json:"results"`
 }
 type RerankUsage struct {
-	TotalTokens int `json:"total_tokens"`
+	TotalTokens int64 `json:"total_tokens"`
 }
 
 func ConvertRerankRequest(meta *meta.Meta, req *http.Request) (string, http.Header, io.Reader, error) {
