+++ relay/model/rerank.go
@@ -21,8 +21,8 @@ type RerankResult struct {
 }
 
 type RerankMetaTokens struct {
-	InputTokens  int `json:"input_tokens"`
-	OutputTokens int `json:"output_tokens"`
+	InputTokens  int64 `json:"input_tokens"`
+	OutputTokens int64 `json:"output_tokens"`
 }
 
 type RerankMeta struct {
